+++ src/lib.rs
@@ -13,112 +13,8 @@ use num::traits::float::FloatCore;
 
 pub mod trigonometry;
 
-pub trait MixedNumConversion<T> {
-    /// Generic type cast from numeric type T.
-    fn mixed_from_num( number:T ) -> Self;
-    /// Generic type cast to  numeric type T.
-    fn mixed_to_num( &self )      -> T;
-}
-
-pub trait MixedTrigonometry
-{
-    /// Take the sin of `self`. Implementation varies with type.
-    fn mixed_sin(&self) -> Self;
-    /// Take the sin of `self`. Implementation varies with type.
-    fn mixed_cos(&self) -> Self;
-    /// Take the sin of `self`. Implementation varies with type.
-    fn mixed_atan(&self) -> Self;
-    /// Take the atan2 of `self`/other. Implementation varies with type.
-    fn mixed_atan2(&self, other:Self) -> Self;
-}
-
-pub trait MixedWrapPhase
-{
-    /// Wrapps `self` to the -π=<x<π range.
-    fn mixed_wrap_phase(&self) -> Self;
-}
-
-pub trait MixedConsts
-{
-    /// The mixed_pi constant. 3.141...
-    fn mixed_pi() -> Self;
-}
-
-pub trait MixedSqrt
-{
-    /// The generic square root implementation for the `MixedSqrt` trait. The implimentation may change with time.
-    /// To stay on this excact implementation, use the `mixed_niirf` function.
-    fn mixed_sqrt(&self) -> Self;
-    /// A fast implementation of the square root using the Nonlinear IIR Filter (NIIRF) method \[1\].
-    /// 
-    /// Only valid for positive values of `self`. Negative values are forced positive before converison.
-    /// Accurate to 5*10⁻⁴ with two iterations \[2\].
-    /// 
-    /// The structure of the estimator is illustrated below \[1\].
-    /// 
-    /// ![Alt version](https://raw.githubusercontent.com/ErikBuer/Fixed-Trigonometry/main/figures/niirf.svg)
-    /// 
-    /// The method utilizes a lookup-table for the acceleration factor β.
-    /// 
-    /// β(x) can be calculated from the following formula, yielding even greater accuracy at a computational cost.
-    /// ```Julia
-    /// β(x) = 0.763x^2-1.5688x+1.314 
-    /// ```
-    /// 
-    /// \[1\] N.Mikami et al., A new DSP-oriented algorithm for calculation of square root using a non-linear digital filter, IEEE Trans. on Signal Processing, July 1992, pp. 1663-1669.
-    /// 
-    /// \[2\] R. G. Lyons, Streamlining Digital Signal Processing, Second Edition, IEEE Press, 2012.
-    /// 
-    /// 
-    /// ## Accuracy and Comparison
-    /// 
-    /// The figure below shows error of the NIIRF implementation, compared to the `std::f32::sqrt` implementation.
-    /// 
-    /// ![Alt version](https://github.com/ErikBuer/Fixed-Trigonometry/blob/main/figures/niirf_sqrt_comparison.png?raw=true)
-    /// 
-    /// Another fixed point implementation of the square root can be found in the cordic crate. 
-    /// 
-    /// Below is the error comparison between the two implementations.
-    /// 
-    /// ![Alt version](https://github.com/ErikBuer/Fixed-Trigonometry/blob/main/figures/sqrt_error_comparison.png?raw=true)
-    fn mixed_niirf(&self) -> Self;
-}
-
-pub trait MixedNum
-    where Self: MixedConsts 
-                + MixedNumConversion<i32> + MixedNumConversion<i64>
-                + MixedNumConversion<f32> + MixedNumConversion<f64>
-                + core::cmp::PartialOrd
-                + core::marker::Sized
-                + core::ops::Div<Output = Self>
-                + core::ops::Sub<Output = Self>
-                + core::ops::Add<Output = Self>
-                + core::ops::Mul<Output = Self>
-                + core::ops::AddAssign
-                + core::ops::SubAssign
-                + Copy
-{
-    /// Maximum value of the type.
-    fn mixed_max_value() -> Self;
-    /// Minimum value of the type.
-    fn mixed_min_value() -> Self;
-    /// Absolute value.
-    fn mixed_abs( &self ) -> Self;
-    /// Integer valued power.
-    fn mixed_powi( &self, exp: i32 ) -> Self;
-    /// Get the sign of the argument with a unit value.
-    /// Zero is of positive sign.
-    fn mixed_sign( &self ) -> Self;
-    /// Returns a bool if self is positive.
-    fn mixed_is_positive( &self ) -> bool;
-    /// Returns a bool if self is negative.
-    fn mixed_is_negative( &self ) -> bool;
-}
-
-pub trait MixedNumSigned
-    where Self: core::ops::Neg<Output = Self>
-{
-}
+pub mod traits;
+use traits::*;
 
 macro_rules! impl_mixed_num_for_primitive{
     ( $T:ty ) => {
@@ -144,6 +40,30 @@ macro_rules! impl_mixed_num_for_primitive{
             #[inline(always)]
             fn mixed_to_num( &self ) -> f64 {
                 return *self as f64;
+            }
+        }
+
+        impl MixedNumConversion<u32> for $T
+        {
+            #[inline(always)]
+            fn mixed_from_num( number:u32 ) -> Self {
+                return number as Self;
+            }
+            #[inline(always)]
+            fn mixed_to_num( &self ) -> u32 {
+                return *self as u32;
+            }
+        }
+
+        impl MixedNumConversion<u64> for $T
+        {
+            #[inline(always)]
+            fn mixed_from_num( number:u64 ) -> Self {
+                return number as Self;
+            }
+            #[inline(always)]
+            fn mixed_to_num( &self ) -> u64 {
+                return *self as u64;
             }
         }
 
@@ -217,10 +137,22 @@ macro_rules! impl_mixed_num_for_primitive{
             }
         }
 
+        impl MixedOps for $T
+        {
+            #[inline(always)]
+            fn mixed_abs( &self ) -> Self {
+                return self.abs();
+            }
+            #[inline(always)]
+            fn mixed_powi( &self, exp: i32 ) -> Self {
+                return self.powi( exp );
+            }
+        }
+
         impl MixedNumSigned for $T
         {   
         }
-
+        
         impl MixedNum for $T
         {
             #[inline(always)]
@@ -230,14 +162,6 @@ macro_rules! impl_mixed_num_for_primitive{
             #[inline(always)]
             fn mixed_min_value() -> Self {
                 return Self::min_value();
-            }
-            #[inline(always)]
-            fn mixed_abs( &self ) -> Self {
-                return self.abs();
-            }
-            #[inline(always)]
-            fn mixed_powi( &self, exp: i32 ) -> Self {
-                return self.powi( exp );
             }
             fn mixed_sign( &self) -> Self {
                 return trigonometry::sign(*self);
@@ -252,12 +176,34 @@ macro_rules! impl_mixed_num_for_primitive{
             }
         }
 
-        impl MixedConsts for $T
+        impl MixedZero for $T
+        {
+            /// Return the zero value of type Self.
+            #[inline(always)]
+            fn mixed_zero() -> Self {
+                return 0 as $T;
+            }
+        }
+
+        impl MixedOne for $T
+        {
+            /// Return the zero value of type Self.
+            #[inline(always)]
+            fn mixed_one() -> Self {
+                return 1 as $T;
+            }
+        }
+
+        impl MixedPi for $T
         {
             #[inline(always)]
             fn mixed_pi() -> Self {
                 return 3.1415926535897932384626433832795028841971693993751058209749445923078164062 as $T;
             }
+        }
+
+        impl MixedConsts for $T
+        {
         }
     }
 }
@@ -292,6 +238,30 @@ macro_rules! impl_mixed_num_for_fixed{
             }
         }
 
+        impl MixedNumConversion<u32> for $T
+        {
+            #[inline(always)]
+            fn mixed_from_num( number:u32 ) -> Self {
+                return Self::from_num(number);
+            }
+            #[inline(always)]
+            fn mixed_to_num( &self ) -> u32 {
+                return self.to_num::<u32>();
+            }
+        }
+
+        impl MixedNumConversion<u64> for $T
+        {
+            #[inline(always)]
+            fn mixed_from_num( number:u64 ) -> Self {
+                return Self::from_num(number);
+            }
+            #[inline(always)]
+            fn mixed_to_num( &self ) -> u64 {
+                return self.to_num::<u64>();
+            }
+        }
+
         impl MixedNumConversion<i32> for $T
         {
             #[inline(always)]
@@ -316,12 +286,34 @@ macro_rules! impl_mixed_num_for_fixed{
             }
         }
 
-        impl MixedConsts for $T
+        impl MixedPi for $T
         {
             #[inline(always)]
             fn mixed_pi() -> Self {
                 return Self::from_num(3.1415926535897932384626433832795028841971693993751058209749445923078164062);
             }
+        }
+
+        impl MixedZero for $T
+        {
+            /// Return the zero value of type Self.
+            #[inline(always)]
+            fn mixed_zero() -> Self {
+                return Self::from_num(0) as $T;
+            }
+        }
+
+        impl MixedOne for $T
+        {
+            /// Return the zero value of type Self.
+            #[inline(always)]
+            fn mixed_one() -> Self {
+                return Self::from_num(1) as $T;
+            }
+        }
+
+        impl MixedConsts for $T
+        {
         }
 
         impl MixedSqrt for $T
@@ -344,6 +336,18 @@ macro_rules! impl_mixed_num_for_fixed_unsigned{
     ( $T:ty ) => {
         impl_mixed_num_for_fixed!($T);
 
+        impl MixedOps for $T
+        {
+            #[inline(always)]
+            fn mixed_abs( &self ) -> Self {
+                return *self; // Is itself for unsigned.
+            }
+            #[inline(always)]
+            fn mixed_powi( &self, exp: i32 ) -> Self {
+                return trigonometry::powi( *self, exp as usize );
+            }
+        }
+
         impl MixedNum for $T
         {
             #[inline(always)]
@@ -353,14 +357,6 @@ macro_rules! impl_mixed_num_for_fixed_unsigned{
             #[inline(always)]
             fn mixed_min_value() -> Self {
                 return Self::MIN;
-            }
-            #[inline(always)]
-            fn mixed_abs( &self ) -> Self {
-                return *self; // Is itself for unsigned.
-            }
-            #[inline(always)]
-            fn mixed_powi( &self, exp: i32 ) -> Self {
-                return trigonometry::powi( *self, exp as usize );
             }
             #[inline(always)]
             fn mixed_sign( &self) -> Self {
@@ -386,6 +382,18 @@ macro_rules! impl_mixed_num_for_fixed_signed{
         {
         }
 
+        impl MixedOps for $T
+        {
+            #[inline(always)]
+            fn mixed_abs( &self ) -> Self {
+                return self.abs();
+            }
+            #[inline(always)]
+            fn mixed_powi( &self, exp: i32 ) -> Self {
+                return trigonometry::powi( *self, exp as usize );
+            }
+        }
+
         impl MixedNum for $T
         {
             #[inline(always)]
@@ -395,14 +403,6 @@ macro_rules! impl_mixed_num_for_fixed_signed{
             #[inline(always)]
             fn mixed_min_value() -> Self {
                 return Self::MIN;
-            }
-            #[inline(always)]
-            fn mixed_abs( &self ) -> Self {
-                return self.abs();
-            }
-            #[inline(always)]
-            fn mixed_powi( &self, exp: i32 ) -> Self {
-                return trigonometry::powi( *self, exp as usize );
             }
             #[inline(always)]
             fn mixed_sign( &self) -> Self {